feat(config): added nestjs/config, joi and config entry points for server config and feature flags
in order to allow programmatic configuration and concentrate dot env access at one point in the
application, nestjs/config has been added and also a place for config files. The main.ts and health
module were updated to use the config service provided by the app in case it is needed. The update
also contains a place for feature flags, should they be necessary.