fix: reverted a dependency update and added an additional step to the pipeline job
an update to a dependency was reverted since it caused npm ci to fail. the check-updates step in the
pipeline got an additional step to verify that the updated package.json is actually installable and
also a conditional setup for the other steps to only persist changes made in the step if the
reinstallation passes.