feat(docker): docker-compose for local development
a docker compose file containing the app and a postgres database have been added to provide a local
dev stack to spin up. Therefore, the docker file multi stage build got updated. The prisma schema
got updated as well to be able to run with certain types of Linux builds and also be compatible with
Apple M1 systems